feat: rest API and websocket (#35)
* chore: add dependencies for springboot to project

* feat: REST controller startGame

* feat: initial websocket setup

* chore: removed unused files

* feat(websocket): wire handleTextMessage, write tests

* refactor: rename cell to tile

Reneamed all instances of cell to tiles.
This adheres to our agreed naming scheme.

* feat(websocket): handle incomming payloads, parse and send snapshot

I did not adhere to the contract proposed by Gateway.
On reflection I think this is a better way of handling the data.
I can change the format, if needed.

* feat: setup springboot, bean config

This completes the feature implimentation.
The maven app can now get launched, and will listen on the assigned
port.

* chore: add authorship

I omitterd authorship on the most simple records. It does not make sense
to add authorship on them.

---------

Co-authored-by: WeihaoMo1 <weihaomo520@gmail.com>