refactor: websocket connection from host to gateway (#67)
* refactor: change websocket from host to gateway

A test is included to verify that the connection has been established correctly.
Another test is included to verify that the message has been sent correctly.

* feat: springboot websocket creation

Springboot never created the new websocket handler. Also connect() was
never called.

* chore: remove file comment

* fix: return same instance gameManager

* refactor: new websocketHandler

We delete the old websocketHandler that waits for connection from gateway, we are now using the new gateway

* fix: remove print statements

* chore: reformat code

---------

Co-authored-by: WeihaoMo1 <weihaomo520@gmail.com>