feat: user story conveyor belts (#164)
* feat: backend logic for blue and green conveyor added

 Two tests are included that show that the conveyor belt pushes the robot in the direction of the arrow and changes direction if the arrival conveyor can rotate the robot.

The next step is to conduct more tests with different scenarios and more board elements, including checkpoints, walls, and blocked robots.

* refactor: Activate board elements based on their priority in PhaseIndex

Phase class has now been expanded with several sub-phases that are triggered in a list.
To avoid adding the same Tile multiple times if it had multiple effects for the same phase we use LinkedHashSet to fix that.

* refactor: a small clean up in enum class Phase

Now we drop the list and use the enum's natural order

* refactor: remove some condition checks

Those checks are always true because of  our phaseIndex, so we remove that

* refactor: remove EnumSet.range

We determine the order in the enum phase.
A test is also included that shows that a robot pushed by another robot on the conveyor belt and the tile effect are still applied in the correct order. The robot is pushed off the conveyor belt and the checkpoint is registered correctly.

* fix: robots on conveyor belts move simultaneously

The first two tests have been modified slightly to show that the conveyor belt is activated at the end of each register.
A test is added to show that a robot has moved off the belt, the conveyor belt no longer affects it, so it moves one space instead of two.

Making robots to move simultaneously is kinda difficult as we can't just looped tiles in board order and moved robots as you found them.
We have to sort the conveyor tiles so that the leading robot is moved first.
It is maybe a bad solution. But I keep it from now on.

* fix: move all robots simultaneously without priority

Now I calculated where every robot intends to go and allow movement into a tile even if another robot is there, as long as that robot will also move. Their positions are updated directly with setter.

Two tests are added to show that robots move simultaneously on blue and green conveyor belts.

* refactor: merge reuse logic

Just realized that I actually have a lot of logic that can be merged....... So I create a class for that purpose.

* feat: collision detection added

We have to process all conveyors at once in stead of process them tile by tile sequentially. It is therefore also need to change applyTileEffect.

Now we predict where all robots will be after each step and if we detect that robots will collide we mark these robots as blocked. Relevant tests are also included.

There are also tests included that say that if a conveyor belt moves a
robot into an area outside the conveyor belt where another robot is located, the robot that is moving must stop at the last area of the conveyor belt. It does not push the robot that is in the way.

Now we only need conveyor belt tests with walls.

* feat: tests with walls and tests with rotation added

Many many test

* feat: client rendering conveyors

We still need a step-by-step execution for conveyor belts so that it doesn't look like the robot is jumping from one place to another.

* refactor: move methods to api

* Revert "Merge branch 'main' into 110-user-story-conveyor-belts"

This reverts commit 75223f3c2e65cdaec9166140ee14fcd33fe2fd62, reversing
changes made to 98d2666493dfdf1432bd742faa986f630f51b62c.

* Reapply "Merge branch 'main' into 110-user-story-conveyor-belts"

This reverts commit d99d939d03ea1e536292fda66cfae2399b9ad678.

* chore: resolve merge conflict

* refactor: small refactor

* refactor: a very nice ui for conveyor belts and rearranging conveyor belt patterns

* fix: image size fixed

* refactor: a huge refactor of conveyor logic

* refactor: allow conveyor belts to be fired on delay

This required a change in the gamescheduler, but now all effects have a
delay.

* chore: authorship added

---------

Co-authored-by: Squirrel Modeller <squirrelmodeller@protonmail.com>